refactor coname and ResultKeyName
