+++ source/databricks/package/constants/__init__.py
@@ -12,16 +12,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from package.constants import Colname
-from pyspark.sql.types import StructType, StructField, StringType, TimestampType
-
-es_brp_relations_schema = StructType(
-    [
-        StructField(Colname.energy_supplier_id, StringType(), False),
-        StructField(Colname.balance_responsible_id, StringType(), False),
-        StructField(Colname.grid_area, StringType(), False),
-        StructField(Colname.metering_point_type, StringType(), False),
-        StructField(Colname.from_date, TimestampType(), False),
-        StructField(Colname.to_date, TimestampType(), False),
-    ]
-)
+from .result_key_name import ResultKeyName
+from .colname import Colname
