+++ source/databricks/package/schemas/integration_event_schema.py
@@ -11,7 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from geh_stream.codelists import Colname
+from package.constants import Colname
 from pyspark.sql.types import StructType, StructField, StringType
 
 integration_event_schema = StructType(
