+++ source/databricks/package/shared/filters.py
@@ -11,7 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from geh_stream.codelists import Colname
+from package.constants import Colname
 from package.shared.period import Period
 from pyspark.sql.dataframe import DataFrame
 from pyspark.sql.functions import col
