+++ source/databricks/package/steps/aggregation/aggregation_result_formatter.py
@@ -15,7 +15,7 @@ from pyspark.sql import DataFrame, SparkSession
 
 from pyspark.sql.functions import lit
 from package.shared.data_classes import Metadata
-from geh_stream.codelists import Colname
+from package.constants import Colname
 from package.schemas.output import aggregation_result_schema
