+++ source/databricks/package/steps/aggregation/combine_master_data.py
@@ -11,7 +11,7 @@
 # # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # # See the License for the specific language governing permissions and
 # # limitations under the License.
-from geh_stream.codelists import Colname, ResultKeyName
+from package.constants import Colname, ResultKeyName
 from pyspark.sql import DataFrame
 from pyspark.sql.functions import col, when
 from package.shared.data_classes import Metadata
