+++ source/databricks/package/steps/wholesale/fee_calculators.py
@@ -13,8 +13,9 @@
 # limitations under the License.
 from pyspark.sql import DataFrame, SparkSession
 from pyspark.sql.functions import col, count, sum
-from geh_stream.codelists import Colname, MarketEvaluationPointType, SettlementMethod
+from geh_stream.codelists import MarketEvaluationPointType, SettlementMethod
 from package.schemas.output import calculate_fee_charge_price_schema
+from package.constants import Colname
 
 
 def calculate_fee_charge_price(
