+++ source/databricks/tests/helpers/test_schemas/tariff_distinct_schema.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from geh_stream.codelists import Colname
+from package.constants import Colname
 from pyspark.sql.types import (
     DecimalType,
     StructType,
