+++ source/databricks/tests/integration/calculator/steps/aggregation/test_combine_master_data.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 from decimal import Decimal
 from datetime import datetime
-from geh_stream.codelists import Colname, ResultKeyName, ResolutionDuration
+from geh_stream.codelists import ResolutionDuration
 from package.steps.aggregation import (
     combine_added_system_correction_with_master_data,
     combine_added_grid_loss_with_master_data,
@@ -31,6 +31,7 @@ import pytest
 import pandas as pd
 from package.schemas.output import aggregation_result_schema
 from tests.helpers import DataframeDefaults
+from package.constants import Colname, ResultKeyName
 
 
 @pytest.fixture(scope="module")
