+++ source/databricks/tests/integration/calculator/steps/aggregation/test_grid_loss_calculation.py
@@ -15,8 +15,6 @@ from decimal import Decimal
 from datetime import datetime, timedelta
 from enum import Enum
 from geh_stream.codelists import (
-    Colname,
-    ResultKeyName,
     MarketEvaluationPointType,
     ResolutionDuration,
     SettlementMethod,
@@ -34,7 +32,7 @@ from pyspark.sql.types import StructType, StringType, DecimalType, TimestampType
 from pyspark.sql.functions import col
 import pytest
 import pandas as pd
-
+from package.constants import Colname, ResultKeyName
 
 date_time_formatting_string = "%Y-%m-%dT%H:%M:%S%z"
 default_obs_time = datetime.strptime(
