+++ source/databricks/tests/integration/calculator/steps/aggregation/test_net_exchange_neighbour.py
@@ -16,7 +16,7 @@ import pytest
 from decimal import Decimal
 import pandas as pd
 from datetime import datetime, timedelta
-from geh_stream.codelists import Colname, ResultKeyName
+from package.constants import Colname, ResultKeyName
 from package.steps.aggregation import (
     aggregate_net_exchange_per_neighbour_ga,
 )
